PhotosController#uploadPhotoObject
* this is mostly the logic moved over from StorageController#upload.
* the upload file is now /userId/fileExtension/fileName.ext
