+++ conf/routes
@@ -1,18 +1,18 @@
-GET /api/ping                   com.authService.controllers.ApiController.ping()
-GET /api/post/:postId           com.authService.controllers.ApiController.getPost(postId: Int)
-GET /api/post/:postId/comments  com.authService.controllers.ApiController.getComments(postId: Int)
-GET /api/addUser                com.authService.controllers.ApiController.addUser()
+GET /api/ping                       com.authService.controllers.ApiController.ping()
+GET /api/post/:postId               com.authService.controllers.ApiController.getPost(postId: Int)
+GET /api/post/:postId/comments      com.authService.controllers.ApiController.getComments(postId: Int)
+GET /api/addUser                    com.authService.controllers.ApiController.addUser()
 
 # Photo routes
-POST /api/photos                 com.authService.controllers.PhotosController.createPhoto()
-GET /api/photos                  com.authService.controllers.PhotosController.getPhotos()
-GET /api/photos/:photoId         com.authService.controllers.PhotosController.getPhoto(photoId: Int)
-PUT /api/photos/:photoId         com.authService.controllers.PhotosController.updatePhoto(photoId: Int)
-DELETE /api/photos/:photoId      com.authService.controllers.PhotosController.deletePhoto(photoId: Int)
+POST /api/photos                    com.authService.controllers.PhotosController.createPhoto()
+GET /api/photos                     com.authService.controllers.PhotosController.getPhotos()
+GET /api/photos/:photoId            com.authService.controllers.PhotosController.getPhoto(photoId: Int)
+PUT /api/photos/:photoId            com.authService.controllers.PhotosController.updatePhoto(photoId: Int)
+DELETE /api/photos/:photoId         com.authService.controllers.PhotosController.deletePhoto(photoId: Int)
+POST /api/photos/:photoId/storage   com.authService.controllers.PhotosController.uploadPhotoObject(photoId: Int)
 
-# Storate routes
-POST /api/storage                com.authService.controllers.StorageController.upload()
-DELETE /api/storage/:fileName    com.authService.controllers.StorageController.delete(fileName: String)
+#Storage routes
+DELETE /api/storage/fileName        com.authService.controllers.StorageController.delete(fileName: String)
 
-GET /auth/login                  com.authService.controllers.AuthenticationController.login()
-POST /auth/register              com.authService.controllers.AuthenticationController.register()
+GET /auth/login                     com.authService.controllers.AuthenticationController.login()
+POST /auth/register                 com.authService.controllers.AuthenticationController.register()
