+++ test/com/authService/controllers/PhotosControllerSpec.scala
@@ -1,19 +1,24 @@
 package com.authService.controllers
 
+import com.amazonaws.services.s3.model.PutObjectResult
 import com.authService.UnitSpec
 import com.authService.auth.{AuthAction, AuthService}
 import com.authService.models.Photo
 import com.authService.repositories.PhotoRepository
+import play.api.libs.Files.TemporaryFile
 import play.api.libs.json.Json
+import play.api.mvc.MultipartFormData.FilePart
 import play.api.mvc._
 import play.api.test._
 import play.api.test.Helpers._
 
 import java.time.Instant
-import scala.util.Success
+import scala.util.{Failure, Success, Try}
 import scala.concurrent.Future
 
 class PhotosControllerSpec extends UnitSpec {
+  import scala.concurrent.ExecutionContext.Implicits.global
+
   val mockPhotoRepository: PhotoRepository = mock[PhotoRepository]
   val controllerComponents: ControllerComponents =
     Helpers.stubControllerComponents()
@@ -335,6 +340,89 @@ class PhotosControllerSpec extends UnitSpec {
           val bodyText: String = contentAsString(response)
           assert(bodyText == s"{\"message\":\"Photo not updated\"}")
         }
+      }
+    }
+  }
+
+  describe("#uploadPhotoObject") {
+    val mockRequest = mock[Request[MultipartFormData[TemporaryFile]]]
+
+    def setupMockRequest(formData: MultipartFormData[TemporaryFile]): Unit = {
+      (mockRequest.headers _).expects().returns(Headers(authHeaders))
+      (mockRequest.body _).expects().returns(formData)
+    }
+
+    def mockUploadObject(returnValue: Try[PutObjectResult]): Unit = {
+      (mockPhotoRepository.uploadObject _)
+        .expects(*, *, *, *)
+        .returns(Future(returnValue))
+    }
+
+    def setupResponse(fileName: String) = {
+      setupAuth()
+
+      val tempFile = play.api.libs.Files.SingletonTemporaryFileCreator
+        .create("mock-photo", "txt")
+      val mockFile =
+        FilePart(fileName, "mock-file", Option("text/plain"), tempFile)
+      val formData = new MultipartFormData(
+        dataParts = Map("" -> Seq("mock_data")),
+        files = Seq(mockFile),
+        badParts = Seq()
+      )
+
+      setupMockRequest(formData)
+
+      controller
+        .uploadPhotoObject(1)
+        .apply(mockRequest)
+    }
+
+    describe("when a valid file is provided") {
+      describe("when the file upload succeeds") {
+        it("should return Ok") {
+          val mockObjectResult = Success(new PutObjectResult())
+          mockUploadObject(mockObjectResult)
+
+          val response = setupResponse("file")
+          assert(status(response) == OK)
+          val bodyText: String = contentAsString(response)
+          assert(bodyText == """{"message":"File uploaded"}""")
+        }
+      }
+
+      describe("when the file upload fails") {
+        it("should return Bad Request") {
+          val mockObjectResult = Failure(new Exception("mock-error"))
+          mockUploadObject(mockObjectResult)
+
+          val response = setupResponse("file")
+          assert(status(response) == BAD_REQUEST)
+          val bodyText: String = contentAsString(response)
+          assert(bodyText == """{"message":"mock-error"}""")
+        }
+      }
+
+      describe("when the file upload errors") {
+        it("should return Bad Request") {
+          (mockPhotoRepository.uploadObject _)
+            .expects(*, *, *, *)
+            .returns(Future.failed(new Exception("mock-error")))
+
+          val response = setupResponse("file")
+          assert(status(response) == BAD_REQUEST)
+          val bodyText: String = contentAsString(response)
+          assert(bodyText == """{"message":"mock-error"}""")
+        }
+      }
+    }
+
+    describe("when a valid file is not provided") {
+      it("should return Bad Request") {
+        val response = setupResponse("not-recognised-file-name")
+        assert(status(response) == BAD_REQUEST)
+        val bodyText: String = contentAsString(response)
+        assert(bodyText == """{"message":"No file found"}""")
       }
     }
   }
